// api/coreService/types/dto/TranslationEditionResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type DtoTranslationEditionResponse = {
    /**
     * @type string | undefined
    */
    author?: string;
    /**
     * @type string | undefined
    */
    createdAt?: string;
    /**
     * @type string | undefined
    */
    createdBy?: string;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type string | undefined
    */
    language?: string;
    /**
     * @type string | undefined
    */
    name?: string;
    /**
     * @type string | undefined
    */
    updatedAt?: string;
    /**
     * @type string | undefined
    */
    updatedBy?: string;
};